Refactored some code from the Transcoder class to a separate Repository class

+++ src/main/java/tech/bitgoblin/config/Config.java
@@ -31,7 +31,7 @@ public class Config {
   }
 
   public int getInt(String key) {
-    return Objects.requireNonNull(this.result.getDouble(key)).intValue();
+    return Objects.requireNonNull(this.result.getLong(key)).intValue();
   }
 
   public boolean contains(String key) {
